Use naudiodon fork, and stop checking for updates when in dev mode
