+++ client/package.json
@@ -26,11 +26,12 @@
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^7.1.9",
+    "electron-is-dev": "^2.0.0",
     "electron-log": "^4.3.1",
     "electron-updater": "^4.3.5",
     "js-base64": "^3.4.4",
     "lodash": "^4.17.19",
-    "naudiodon": "^2.3.4",
+    "naudiodon": "https://github.com/gasnew/naudiodon",
     "react": "^16.13.1",
     "react-beautiful-dnd": "^13.0.0",
     "react-dom": "^16.13.1",
